Fix use of size_t to represent type ID

+++ lib/mlir/Dialect/Clift/IR/CliftOps.cpp
@@ -189,7 +189,7 @@ struct ModuleValidator {
 private:
   llvm::SmallPtrSet<mlir::Type, 2> VisistedTypes;
   llvm::SmallPtrSet<mlir::Attribute, 2> VisitedAttrs;
-  llvm::DenseMap<size_t, mlir::clift::TypeDefinitionAttr> Definitions;
+  llvm::DenseMap<uint64_t, mlir::clift::TypeDefinitionAttr> Definitions;
 };
 
 mlir::LogicalResult mlir::clift::ModuleOp::verify() {
